% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{as_handler}
\alias{as_handler}
\title{As Handler}
\usage{
as_handler(f, conditions)
}
\arguments{
\item{f}{a function to be a condtionr handler object}
}
\description{
Cast a function to be a condtionr handler object
}
