% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{condition_factory}
\alias{condition_factory}
\title{Create a function to be used to create new condition signals}
\usage{
condition_factory(type)
}
\arguments{
\item{type}{character(n) condition signal type of the resultant function}
}
\description{
Creates a function which can be used to create new condition signals of
a given type.
}
