% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextualize.R
\name{contextualize}
\alias{contextualize}
\title{Contextualize (Simple) Conditions}
\usage{
contextualize(expr, ..., regex = FALSE, include_original_message = TRUE)
}
\arguments{
\item{expr}{an expression which potentially signals a condition.}

\item{...}{(named) lists as named arguments. The elements of one of these
lists are passed as arguments to \code{stackCondition}, i.e. an element
"message" must be present if \code{include_original_message == FALSE}. If
missing, "type" is automatically derived from the type of the caught
condition. "call" is ignored.}

\item{regex}{a logical value. If \code{FALSE} (the default), then the names
of the \code{...} arguments refer to the class of the condition object. If
\code{TRUE}, the names of the \code{...} arguments are treated as regular
expressions against which the condition message is matched.}

\item{include_original_message}{a logical value. Should the message of the
caught condition be appended to the message specified in the chosen
\code{...} list?}
}
\value{
whatever \code{expr} returns.
}
\description{
\code{contextualize} catches (simple) conditions and instead signals a
stacked condition.
}

