% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exprToQuo}
\alias{exprToQuo}
\title{expression to quosure}
\usage{
exprToQuo(expr, env = parent.frame(), quoted = FALSE)
}
\description{
This is almost verbatim a copied implementation of this internal shiny function.
Prevents us from relying on an unexported function from that package.
}
\keyword{internal}
