% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quoToLabelBody}
\alias{quoToLabelBody}
\title{quosure to label body}
\usage{
quoToLabelBody(q)
}
\description{
This is almost verbatim a copied implementation of this internal shiny function.
Prevents us from relying on an unexported function from that package.
}
\keyword{internal}
