% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_shiny_condition_handler}
\alias{register_shiny_condition_handler}
\title{Register a condition hander into the set of defaults}
\usage{
register_shiny_condition_handler(f)
}
\arguments{
\item{f}{a function of class conditionr_handler}
}
\description{
Register a condition hander into the set of defaults
}
