% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_class}
\alias{with_class}
\title{With class}
\usage{
with_class(x, cls)
}
\arguments{
\item{x}{an object}

\item{cls}{character(n) classes to attach to x}
}
\value{
x with additional classes
}
\description{
A convenience method for attaching new classes to a given object
}
\keyword{internal}
