% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextualize.R
\name{contextualize}
\alias{contextualize}
\title{Contextualize (Simple) Conditions}
\usage{
contextualize(expr, ..., include_original_message = TRUE)
}
\arguments{
\item{expr}{an expression which potentially signals a condition.}

\item{...}{(named) lists as named arguments. The elements of one of these
lists are passed as arguments to \code{stackCondition}, i.e. an element
"message" must be present if \code{include_original_message == FALSE}. If
missing, "type" is automatically derived from the type of the caught
condition. "call" is ignored. The names of these arguments refer to the
class of the condition object.}

\item{include_original_message}{a logical value. Should the message of the
caught condition be appended to the message specified in the chosen
\code{...} list?}
}
\value{
whatever \code{expr} returns.
}
\description{
\code{contextualize} catches (simple) conditions and instead signals a
stacked condition.
}

